+++ cmd/complyctl/cli/options.go
@@ -22,7 +22,7 @@ type Output struct {
 }
 
 func (o *Common) BindFlags(fs *pflag.FlagSet) {
-	fs.BoolVarP(&o.Debug, "debug", "d", false, "output debug logs")
+	fs.BoolVarP(&o.Debug, "debug", "d", false, "output debug logs to stderr and log file")
 	fs.StringVarP(&o.Workspace, "workspace", "w", "", "workspace directory (env: COMPLYTIME_WORKSPACE, default: current directory)")
 }